Apply some positioning to the cards
It is not how I want yet.

I think I'll introduce several variables:

Major x (which of the 7 major x positions the card is in).
Major y (if the card is on the top row or botttom row).
z (how high the card is- visually this will make a small addition to
the x and or y values)
Tableau face down and tableau face up (how many face down and face up
cards are below it on the tableau stack). Visually this will make a
medium addition to the height, where facedown cards are given slightly
less height than faceup cards.